Switch LabelChip to typescript
