+++ src/LabelSelector/LabelChip/LabelChip.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 
-import LabelChip from "./";
+import LabelChip from ".";
 import React from "react";
 import userEvent from "@testing-library/user-event";
