+++ src/LabelSelector/LabelChip/LabelChip.types.ts
@@ -1,0 +1,6 @@
+import { ChipProps } from "@mui/material";
+
+export interface LabelChipProps extends Omit<ChipProps, "color" | "label"> {
+  color?: string;
+  label: string;
+}
