+++ src/LabelSelector/LabelChip/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./LabelChip";
+export { LabelChipProps } from "./LabelChip.types";
