+++ src/LabelSelector/index.ts
@@ -1,3 +1,3 @@
 export { default } from "./LabelSelector";
 export { default as LabelSelector } from "./LabelSelector";
-export { default as LabelChip } from "./LabelChip";
+export { default as LabelChip, LabelChipProps } from "./LabelChip";
