+++ src/index.ts
@@ -26,7 +26,7 @@ export {
 export { default as FontPicker } from "./FontPicker";
 export { default as FontStyle } from "./FontStyle";
 export { default as ImageUploader } from "./ImageUploader";
-export { LabelChip, LabelSelector } from "./LabelSelector";
+export { LabelChip, LabelChipProps, LabelSelector } from "./LabelSelector";
 export { default as Loading } from "./Loading";
 export { default as LoginForm } from "./LoginForm";
 export { default as ModelButton } from "./ModelButton";
